+++ packages/@react-elf/alert/vite.dist.config.ts
@@ -29,12 +29,15 @@ export default defineConfig({
     rollupOptions: {
       // make sure to externalize deps that shouldn't be bundled
       // into your library
-      external: ["react"],
+      external: ["react", "react-dom", "classnames", "@react-elf/shared"],
       output: {
         // Provide global variables to use in the UMD build
         // for externalized deps
         globals: {
           react: "React",
+          "react-dom": "ReactDOM",
+          classnames: "classNames",
+          "@react-elf/shared": "ReactElfShared",
         },
       },
     },