+++ packages/@react-elf/button/vite.dist.config.js
@@ -29,12 +29,20 @@ export default defineConfig({
     rollupOptions: {
       // make sure to externalize deps that shouldn't be bundled
       // into your library
-      external: ["react"],
+      external: [
+        "react",
+        "@react-elf/shared",
+        "@react-elf/tooltip",
+        "classnames"
+      ],
       output: {
         // Provide global variables to use in the UMD build
         // for externalized deps
         globals: {
           react: "React",
+          "@react-elf/shared": "ReactElfShared",
+          "@react-elf/tooltip": "ReactElfTooltip",
+          classnames: "classNames"
         },
       },
     },