+++ packages/@react-elf/layout/vite.dist.config.ts
@@ -29,12 +29,14 @@ export default defineConfig({
     rollupOptions: {
       // make sure to externalize deps that shouldn't be bundled
       // into your library
-      external: ["react"],
+      external: ["react", "@react-elf/shared", "classnames"],
       output: {
         // Provide global variables to use in the UMD build
         // for externalized deps
         globals: {
           react: "React",
+          "@react-elf/shared": "ReactElfShared",
+          classnames: "classNames",
         },
       },
     },