+++ packages/@react-elf/tooltip/vite.dist.config.ts
@@ -29,13 +29,15 @@ export default defineConfig({
     rollupOptions: {
       // make sure to externalize deps that shouldn't be bundled
       // into your library
-      external: ["react", "react-dom"],
+      external: ["react", "react-dom", "@react-elf/shared", "classnames"],
       output: {
         // Provide global variables to use in the UMD build
         // for externalized deps
         globals: {
           react: "React",
           "react-dom": "ReactDOM",
+          "@react-elf/shared": "ReactElfShared",
+          classnames: "classNames",
         },
       },
     },