+++ packages/@react-elf/dialog/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @react-elf/dialog
 
+## 0.0.37
+
+### Patch Changes
+
+- modify vite config
+
+- Updated dependencies []:
+  - @react-elf-types/dialog@0.0.37
+  - @react-elf/button@0.0.37
+
+## 0.0.36
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @react-elf-types/dialog@0.0.36
+  - @react-elf/button@0.0.36
+
 ## 0.0.35
 
 ### Patch Changes