+++ packages/@react-elf/menu/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @react-elf/button
 
+## 0.0.37
+
+### Patch Changes
+
+- modify vite config
+
+- Updated dependencies []:
+  - @react-elf-types/menu@0.0.37
+
+## 0.0.36
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @react-elf-types/menu@0.0.36
+
 ## 0.0.35
 
 ### Patch Changes